LayoutTests/crypto/subtle/hmac-import-raw-key-export-jwk-key-expected.txt

Test importing a raw HMAC key with SHA-1 and then export it in JWK format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
Exporting a key...
PASS Base64URL.parse(exportedJwkKey.k) is rawKey
PASS successfullyParsed is true

TEST COMPLETE
